Module Bechamel_perf

This module provides several event counters from the perf API provided by Linux (2.6+) - and they are not compatible with other systems. These counters are basically what the perf command provides.

For some events, it is necessary to be root. If you try to use these counters with insufficient privileges, Bechamel_perf will report you an permission error.

If you want to use Bechamel_perf as a normal user, you should take a look on the documentation about perf_paranoid.

module Extension : sig ... end

module Instance : sig ... end